-Phan quyen

+++ src/main/java/vn/edu/hcmuaf/ttt/admin/controler/DeleteProduct.java
@@ -1,5 +1,6 @@
 package vn.edu.hcmuaf.ttt.admin.controler;
 
+import vn.edu.hcmuaf.ttt.bean.User;
 import vn.edu.hcmuaf.ttt.service.ProductService;
 
 import javax.servlet.*;
@@ -11,9 +12,17 @@ import java.io.IOException;
 public class DeleteProduct extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String id = request.getParameter("id");
-        ProductService.deleteProduct(id);
-        response.sendRedirect("ProAdmin");
+        User user = (User) request.getSession().getAttribute("auth");
+        boolean isLoggedIn = user != null;
+        boolean isNormalUser = isLoggedIn && user.getUser_admin() != 1;
+        if (!isLoggedIn || isNormalUser)  {
+            response.sendRedirect("/THDoAn_war/List-Product");
+        } else {
+            String id = request.getParameter("id");
+            ProductService.deleteProduct(id);
+            response.sendRedirect("ProAdmin");
+        }
+
     }
 
     @Override
