+++ src/main/java/vn/edu/hcmuaf/ttt/admin/controler/DeleteUser.java
@@ -1,6 +1,7 @@
 package vn.edu.hcmuaf.ttt.admin.controler;
 
 import vn.edu.hcmuaf.ttt.admin.service.UserService;
+import vn.edu.hcmuaf.ttt.bean.User;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
@@ -11,9 +12,17 @@ import java.io.IOException;
 public class DeleteUser extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String id = request.getParameter("id");
-        UserService.deleteUser(id);
-        response.sendRedirect("UserList");
+        User user = (User) request.getSession().getAttribute("auth");
+        boolean isLoggedIn = user != null;
+        boolean isNormalUser = isLoggedIn && user.getUser_admin() != 1;
+        if (!isLoggedIn || isNormalUser)  {
+            response.sendRedirect("/THDoAn_war/List-Product");
+        } else {
+            String id = request.getParameter("id");
+            UserService.deleteUser(id);
+            response.sendRedirect("UserList");
+        }
+
     }
 
     @Override
